+++ packages/terra-framework-docs/src/terra-dev-site/test/slide-panel-manager/SlidePanelManagerWithDisclosureAccessory.test.jsx
@@ -1,17 +1,19 @@
 import React from 'react';
 import classNames from 'classnames/bind';
+import SlidePanelManager from 'terra-slide-panel-manager';
 import DisclosureComponent from './DisclosureComponent';
-import SlidePanelManager from '../../../SlidePanelManager';
 import styles from './SlidePanelManagerDefault.test.module.scss';
 
 const cx = classNames.bind(styles);
 
-const SlidePanelManagerDefault = () => (
+const SlidePanelManagerWithDisclosureAccessory = () => (
   <div role="main" className={cx('content-wrapper')}>
-    <SlidePanelManager>
+    <SlidePanelManager
+      disclosureAccessory={<div className={cx('disclosure-accessory')}>Disclosure Accessory</div>}
+    >
       <DisclosureComponent identifier="root-component" disclosureType="panel" renderHeaderAdapter />
     </SlidePanelManager>
   </div>
 );
 
-export default SlidePanelManagerDefault;
+export default SlidePanelManagerWithDisclosureAccessory;