+++ packages/terra-framework-docs/CHANGELOG.md
@@ -18,6 +18,8 @@
   * Migrate docs from terra-theme-provider
   * Migrate docs from terra-theme-context
   * Migrate docs from terra-tabs
+  * Migrate docs from terra-slide-panel-manager
+
 
 ## 1.7.0 - (March 8, 2022)